fix: moved is_es and is_ec
